requirements.txt
flask
flask-cors
pillow
pytesseract
requests